changed tables

+++ db/migrate/20181120210901_create_songs.rb
@@ -3,7 +3,7 @@ class CreateSongs < ActiveRecord::Migration[5.2]
     create_table :songs do |t|
       t.string :title
       t.string :artist
-      t.integer :bpm
+      t.float :bpm
       t.string :url
       t.string :tags
